Registering UserStore in constructor instead of StoreConfig
